YAYYYYY!!!!! CAN DIE BY TOUCHING ENEMY!!! ENEMY DIE AFTER TOUCHING TRAPS ARND 3 TIMES
Certain problems: I set the enemy damage to 30. If it catches u, u are
almost surely die. There is also a problem with
GameObject.prototype.equals in GameObject line 210. I think it is about
error checking, but it is causing error. As such, I removed the
functions using them: GameEngine line 578 and 591 - 593.