Background Change and BUG REPORT
Since the background was quite dull, I tried to make another one. I kept
the other one as a copy. You could see it. And I know the bug already,
but dpn't have the enrgy to debug. It is the traps; although they are
not activated yet and although they are unseen, they still give damage
to the enemies.