Regular Update
Hero Sprite is done, so I put it in. I increased the sprite contrast
though, because the colour is too light compared to the enemies. I put
the path as not rendered, to see the game in reality. I reduced the
damage from the enemies to make it more reasonable. I guess we really
need the blood explosion, because the enemies now just gone.